backend/requirements.txt
python-dotenv==1.0.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter>=3.5.0
psycopg2-binary==2.9.11
redis>=4.5.2,<5.0.0
celery[redis]==5.3.4
marshmallow==3.20.1
reportlab==4.0.7
gunicorn==21.2.0
